requirements.txt
# Skyy Facial Recognition MCP Server Dependencies
# Versions based on working development environment

# MCP Server Framework
fastmcp==2.13.0.2
mcp==1.20.0

# Facial Recognition
# IMPORTANT FOR WINDOWS: InsightFace 0.7.3 has NO binary on PyPI
# You MUST use the included wheel file for Windows without C++ compiler:
#   pip install insightface-0.7.3-cp311-cp311-win_amd64.whl
#
# For Linux/macOS or Windows with C++ build tools:
#   insightface==0.7.3
# (Uncomment the line below only if you have C++ build tools installed)
# insightface==0.7.3

# Deep Learning - MUST BE FIRST
# CRITICAL: Install NumPy BEFORE opencv-python to prevent version conflicts
# InsightFace 0.7.3 wheel requires NumPy 1.x
numpy==1.26.4
# Use version ranges to allow pip to find compatible versions
onnx>=1.15.0,<1.20.0
onnxruntime>=1.16.0,<1.24.0

# Image Processing
# Allow pip to find opencv-python version compatible with numpy 1.26.4
opencv-python>=4.8.0,<4.13.0
opencv-python-headless>=4.8.0,<4.13.0
Pillow>=10.0.0,<=12.0.0

# InsightFace Dependencies
# Required by InsightFace but not installed with --no-deps
tqdm>=4.66.0
matplotlib>=3.8.0
scipy>=1.13.0
scikit-learn>=1.5.0
scikit-image>=0.24.0
cython>=3.0.0
easydict>=1.13
prettytable>=3.0.0
albumentations>=1.3.0

# Data Validation
pydantic==2.12.3
pydantic-settings==2.11.0

# OAuth 2.1 and Security
PyJWT>=2.8.0
cryptography>=41.0.0
python-multipart>=0.0.6
